pyproject.toml
[tool.poetry]
name = "open-sea-api-client"
version = "2.0.0"
description = "A client library for accessing OpenSea API"
authors = []
readme = "README.md"
packages = [
    {include = "open_sea_api_client"},
]
include = ["CHANGELOG.md", "open_sea_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
